chore: rename component to better reflect purpose
